backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
torch==2.4.1
numpy==1.26.4
transformers==4.45.2
faiss-cpu==1.8.0
rdkit==2024.3.5
httpx==0.27.2
